requirements.txt
alembic==1.16.5
blinker==1.9.0
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.1.8
cloudpickle==3.1.1
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
docker==7.1.0
entrypoints==0.4
Flask==3.1.2
fonttools==4.60.1
gitdb==4.0.12
GitPython==3.1.45
graphene==3.4.3
graphql-core==3.2.6
graphql-relay==3.2.0
greenlet==3.2.4
idna==3.11
importlib_metadata==7.2.1
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.2
kiwisolver==1.4.7
Mako==1.3.10
Markdown==3.9
MarkupSafe==3.0.3
matplotlib==3.9.4
mlflow==2.11.0
numpy==1.26.4
packaging==23.2
pandas==2.3.3
pillow==11.3.0
protobuf==4.25.8
pyarrow==15.0.2
pyparsing==3.2.5
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
pytz==2024.2
pywin32==311; platform_system == "Windows"
PyYAML==6.0.3
querystring-parser==1.2.4
requests==2.32.5
scikit-learn==1.6.1
scipy==1.13.1
six==1.17.0
smmap==5.0.2
SQLAlchemy==2.0.44
sqlparse==0.5.3
threadpoolctl==3.6.0
tomli==2.3.0
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
waitress==3.0.2
Werkzeug==3.1.3
zipp==3.23.0
